Other: Renamed the ClassicEditor file to "classiceditor.js" to match the naming convention.
